Add fix to remove unused corrupt elements
In vanilla RCT2, corrupt elements skips the painting of elements later in the list. In OpenRCT2 only the next element will be skipped. In both cases, they serve no purpose when there are no elements after it, and will only cause objects to seemingly randomly not appear when built. This function loops over all tiles on the map and removes corrupt elements at the end of the list.